fix: Complete C# to Metal kernel translation implementation
- Fixed compilation errors in CSharpToMetalTranslator
- Resolved circular dependency between MetalAccelerator and MetalMemoryManager
- Added weak reference pattern for accelerator reference
- Fixed MetalKernelCompiler to accept MSL directly
- Added buffer view support for kernel arguments
- Created comprehensive translation tests
- Fixed type casting issues (long to nuint)
- Removed duplicate SymbolExtensions class
- Fixed IsReadOnly property access on ISymbol

This completes the C# to Metal kernel translation feature, enabling
developers to write kernels once in C# with [Kernel] attribute and
have them automatically translated to Metal Shading Language.

+++ src/Backends/DotCompute.Backends.Metal/Kernels/MetalCompiledKernel.cs
@@ -163,7 +163,7 @@ MetalCommandBufferPool? commandBufferPool = null) : ICompiledKernel
                     // Handle buffer view with proper offset
                     var parentBuffer = view.ParentBuffer;
                     var offsetBytes = view.Offset;
-                    MetalNative.SetBuffer(encoder, parentBuffer, offsetBytes, bufferIndex);
+                    MetalNative.SetBuffer(encoder, parentBuffer, (nuint)offsetBytes, bufferIndex);
                 }
                 else
                 {
